feat: :construction: fetch `fetcher` and better use of post state

+++ src/routes/api/v1/post/[id]/unlike/+server.ts
@@ -8,8 +8,16 @@ export const POST: RequestHandler = async (event) => {
 
 	const { id: postId } = event.params;
 	try {
-		// should we block unlikes if the author's profile is private
-		// should we block unlikes if the post is marked as deleted
+		// check if the author's profile is private or the post is deleted
+		const isValid = !(await prisma.post.findFirst({
+			select: { id: true },
+			where: {
+				id: postId,
+				OR: [{ author: { privacySettings: { private: true } } }, { deletedAt: { not: null } }]
+			}
+		}));
+		if (!isValid) throw 0;
+
 		const result = await prisma.$transaction([
 			prisma.like.delete({
 				where: {
